Centralise functional test classes
Which provides better isolation and makes things a bit more
understandable.
