--- tests/Doctrine/Tests/DbalTypes/CustomIdObjectHashType.php
@@ -1,45 +1,0 @@
-<?php
-
-namespace Doctrine\Tests\DbalTypes;
-
-use Doctrine\DBAL\Platforms\AbstractPlatform;
-use Doctrine\DBAL\Types\Type;
-
-class CustomIdObjectHashType extends Type
-{
-    const NAME = 'CustomIdObjectHash';
-
-    /**
-     * {@inheritdoc}
-     */
-    public function convertToDatabaseValue($value, AbstractPlatform $platform)
-    {
-        return $value->id.'_test';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function convertToPHPValue($value, AbstractPlatform $platform)
-    {
-        $idObject = new CustomIdObject(str_replace('_test', '', $value));
-
-        return $idObject;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getSQLDeclaration(array $fieldDeclaration, AbstractPlatform $platform)
-    {
-        return $platform->getVarcharTypeDeclarationSQL($fieldDeclaration);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getName()
-    {
-        return self::NAME;
-    }
-}
